Add appropriate new documentation.

+++ docs/docs.py
@@ -1,7 +1,20 @@
 
+import argparse
 import os
 import shutil
 import subprocess
+
+parser = argparse.ArgumentParser()
+parser.add_argument('--serve', action='store_true')
+parser.add_argument('--deploy', action='store_true')
+args = parser.parse_args()
+
+if not args.serve and not args.deploy:
+    print("Need at least one of serve or deploy")
+    parser.print_help()
+    exit(2);
+
+
 
 # obliterate caches, just for safety's sake
 if os.path.exists("docs/obj"): shutil.rmtree("docs/obj")
@@ -31,34 +44,35 @@ subprocess.check_call([
         "docs/docfx.json",
     ])
 
-if False:
+if args.deploy:
+    # update repo
+    os.chdir("docs/_site")
+
+    subprocess.check_call([
+            "git",
+            "add", ".",
+        ])
+
+    subprocess.check_call([
+            "git",
+            "commit",
+            "-a",
+            "-m", "Update documentation.",
+        ])
+
+    subprocess.check_call([
+            "git",
+            "push",
+        ])
+        
+    # ugh this is so ugly
+    os.chdir("../..");
+
+if args.serve:
     subprocess.check_call([
             "docfx",
             "serve", "docs/_site",
         ])
-
-# update repo
-os.chdir("docs/_site")
-
-subprocess.check_call([
-        "git",
-        "add", ".",
-    ])
-
-subprocess.check_call([
-        "git",
-        "commit",
-        "-a",
-        "-m", "Update documentation.",
-    ])
-
-subprocess.check_call([
-        "git",
-        "push",
-    ])
-    
-# ugh this is so ugly
-os.chdir("../..");
-
+        
 # clear repo directory
 shutil.rmtree("docs/_site")
